Move all stories next to component
