+++ src/components/Price/Price.story.js
@@ -3,8 +3,8 @@ import { flow } from 'lodash/fp';
 import { storiesOf } from '@storybook/react';
 import { withInfo } from '@storybook/addon-info';
 
-import withTests from '../util/withTests';
-import { Price } from '../../src/components/Price';
+import withTests from '../../util/withTests';
+import { Price } from './Price';
 
 // storiesOf('Prices', module)
 //   .addDecorator(withTests('Price'))
